+++ packages/headless/src/controllers/breadcrumb-manager/headless-breadcrumb-manager.ts
@@ -7,6 +7,7 @@ import {logCategoryFacetBreadcrumb} from '../../features/facets/category-facet-s
 import {categoryFacetResponseSelectedValuesSelector} from '../../features/facets/category-facet-set/category-facet-set-selectors';
 import {categoryFacetSetReducer as categoryFacetSet} from '../../features/facets/category-facet-set/category-facet-set-slice';
 import {
+  toggleExcludeFacetValue,
   toggleSelectFacetValue,
   updateFreezeCurrentValues,
 } from '../../features/facets/facet-set/facet-set-actions';
@@ -14,19 +15,29 @@ import {logFacetBreadcrumb} from '../../features/facets/facet-set/facet-set-anal
 import {facetResponseActiveValuesSelector} from '../../features/facets/facet-set/facet-set-selectors';
 import {facetSetReducer as facetSet} from '../../features/facets/facet-set/facet-set-slice';
 import {FacetValue} from '../../features/facets/facet-set/interfaces/response';
+import {FacetSlice} from '../../features/facets/facet-set/facet-set-state';
 import {logClearBreadcrumbs} from '../../features/facets/generic/facet-generic-analytics-actions';
-import {toggleSelectDateFacetValue} from '../../features/facets/range-facets/date-facet-set/date-facet-actions';
+import {
+  toggleExcludeDateFacetValue,
+  toggleSelectDateFacetValue,
+} from '../../features/facets/range-facets/date-facet-set/date-facet-actions';
 import {logDateFacetBreadcrumb} from '../../features/facets/range-facets/date-facet-set/date-facet-analytics-actions';
-import {dateFacetSelectedValuesSelector} from '../../features/facets/range-facets/date-facet-set/date-facet-selectors';
+import {dateFacetActiveValuesSelector} from '../../features/facets/range-facets/date-facet-set/date-facet-selectors';
 import {dateFacetSetReducer as dateFacetSet} from '../../features/facets/range-facets/date-facet-set/date-facet-set-slice';
-import {toggleSelectNumericFacetValue} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
+import {DateFacetSlice} from '../../features/facets/range-facets/date-facet-set/date-facet-set-state';
+import {
+  toggleExcludeNumericFacetValue,
+  toggleSelectNumericFacetValue,
+} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
 import {logNumericFacetBreadcrumb} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-analytics-actions';
-import {numericFacetSelectedValuesSelector} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-selectors';
+import {numericFacetActiveValuesSelector} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-selectors';
 import {numericFacetSetReducer as numericFacetSet} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice';
+import {NumericFacetSlice} from '../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-state';
 import {executeSearch} from '../../features/search/search-actions';
 import {searchReducer as search} from '../../features/search/search-slice';
 import {
   logStaticFilterDeselect,
+  toggleExcludeStaticFilterValue,
   toggleSelectStaticFilterValue,
 } from '../../features/static-filter-set/static-filter-set-actions';
 import {
@@ -53,6 +64,7 @@ import {
   DeselectableValue,
   FacetBreadcrumb,
   getBreadcrumbs,
+  GetBreadcrumbsConfiguration,
   NumericFacetBreadcrumb,
   StaticFilterBreadcrumb,
 } from '../core/breadcrumb-manager/headless-core-breadcrumb-manager';
@@ -114,10 +126,10 @@ export function buildBreadcrumbManager(
   const getState = () => engine.state;
 
   const getFacetBreadcrumbs = (): FacetBreadcrumb[] => {
-    return getBreadcrumbs(
+    const config: GetBreadcrumbsConfiguration<Record<string, FacetSlice>> = {
       engine,
-      getState().facetSet,
-      ({facetId, selection}) => {
+      facetSet: getState().facetSet,
+      executeToggleSelect: ({facetId, selection}) => {
         const analyticsAction = logFacetBreadcrumb({
           facetId: facetId,
           facetValue: selection.value,
@@ -128,32 +140,59 @@ export function buildBreadcrumbManager(
         );
         dispatch(executeSearch(analyticsAction));
       },
-      facetResponseActiveValuesSelector
-    );
+      executeToggleExclude: ({facetId, selection}) => {
+        const analyticsAction = logFacetBreadcrumb({
+          facetId: facetId,
+          facetValue: selection.value,
+        });
+        dispatch(toggleExcludeFacetValue({facetId, selection}));
+        dispatch(
+          updateFreezeCurrentValues({facetId, freezeCurrentValues: false})
+        );
+        dispatch(executeSearch(analyticsAction));
+      },
+      facetValuesSelector: facetResponseActiveValuesSelector,
+    };
+
+    return getBreadcrumbs(config);
   };
 
   const getNumericFacetBreadcrumbs = (): NumericFacetBreadcrumb[] => {
-    return getBreadcrumbs(
+    const config: GetBreadcrumbsConfiguration<
+      Record<string, NumericFacetSlice>
+    > = {
       engine,
-      getState().numericFacetSet,
-      (payload) => {
+      facetSet: getState().numericFacetSet,
+      executeToggleSelect: (payload) => {
         dispatch(toggleSelectNumericFacetValue(payload));
         dispatch(executeSearch(logNumericFacetBreadcrumb(payload)));
       },
-      numericFacetSelectedValuesSelector
-    );
+      executeToggleExclude: (payload) => {
+        dispatch(toggleExcludeNumericFacetValue(payload));
+        dispatch(executeSearch(logNumericFacetBreadcrumb(payload)));
+      },
+      facetValuesSelector: numericFacetActiveValuesSelector,
+    };
+    return getBreadcrumbs(config);
   };
 
   const getDateFacetBreadcrumbs = (): DateFacetBreadcrumb[] => {
-    return getBreadcrumbs(
-      engine,
-      getState().dateFacetSet,
-      (payload) => {
-        dispatch(toggleSelectDateFacetValue(payload));
-        dispatch(executeSearch(logDateFacetBreadcrumb(payload)));
-      },
-      dateFacetSelectedValuesSelector
-    );
+    const config: GetBreadcrumbsConfiguration<Record<string, DateFacetSlice>> =
+      {
+        engine,
+        facetSet: getState().dateFacetSet,
+        executeToggleSelect: (payload) => {
+          dispatch(toggleSelectDateFacetValue(payload));
+          dispatch(executeSearch(logDateFacetBreadcrumb(payload)));
+        },
+        executeToggleExclude: (payload) => {
+          dispatch(toggleExcludeDateFacetValue(payload));
+          dispatch(executeSearch(logDateFacetBreadcrumb(payload)));
+        },
+        facetValuesSelector: dateFacetActiveValuesSelector,
+      };
+
+    return getBreadcrumbs(config);
   };
 
   const getCategoryFacetBreadcrumbs = (): CategoryFacetBreadcrumb[] => {
@@ -197,7 +236,7 @@ export function buildBreadcrumbManager(
   ): StaticFilterBreadcrumb => {
     const {id, values: filterValues} = filter;
     const values = filterValues
-      .filter((value) => value.state === 'selected')
+      .filter((value) => value.state !== 'idle')
       .map((value) => buildStaticFilterBreadcrumbValue(id, value));
 
     return {id, values};
@@ -216,7 +255,11 @@ export function buildBreadcrumbManager(
           staticFilterValue: {caption, expression},
         });
 
-        dispatch(toggleSelectStaticFilterValue({id, value}));
+        if (value.state === 'selected') {
+          dispatch(toggleSelectStaticFilterValue({id, value}));
+        } else if (value.state === 'excluded') {
+          dispatch(toggleExcludeStaticFilterValue({id, value}));
+        }
         dispatch(executeSearch(analytics));
       },
     };